plugin: optimize
Stopping the server takes about 0.5 seconds at worst case, which is not
good.

So, from now we will start the server on the first invocation of the plugin.
And stop it when the session finished.

Which means: one server for all your test! :)

One thing we need to ensure: the server is cleared properly between test
runs.