server: implement 'permanently failed' state
This states that at least one request did not meet the requiremenets for the
ordered handlers.

Eg. we waited for a GET /example request to be the *first* request, but the
client sent a GET /foobar request, then we should pull the handbrake and
stop serving further requests.

The idea behing this is to avoid the situation having the 2nd ordered
expectation pass while the 1st failed.