README: convert from rst to md
pypi supports markdown from 2018 and it is better supported on github
Also, it is easier to write.

This commit makes no changes in the contents, only the format is changed
with the aim to have an identical output on github for markdown.

MANIFEST.in needed to be added because 'setup.py sdist' didn't include
README.md in the source distribution.