httpserver: fix query_string type
In werzeug the requets.query_string is a bytes object so expecting a string
in the RequestMatcher breaks the query_string matching functionality. Users
can provide a bytes object but in this case they will break the official
API.

In order to not break the API, an implicit conversion is added which
converts the string object to bytes. Accepting bytes is also added in case
the user does not want this implicit behavior.

Fixes #16