httpserver: add query matcher feature
If query_string is specified as a Mapping (dict), the query string of the
request will be parsed and will be matched with the dictionary specified.
When a key has multiple values in the request (eg. "foo=bar&foo=baz"), the
first item will be used.

If it is desired to match all the values from the same key, use
werkzeug.datastructures.MultiDict instance, as it can hold multiple values
for the same key.