README.md: add donate button (fixes #29)
This is still (and will be) a free and open source project. I don't want to
monetize it, just make the possibility to anyone who willing to donate a
small amount of money.

I'm doing it in my freetime and I'm dedicated to give this piece of software
to everyone finding it useful, free of charge.

Currently, this project is based heavily on werkzeug. Without them,
implementation would not have been smooth and stable. I'm adding their
donate link as well.