poetry: use poetry instead of setuptools
Time to move forward and use poetry instead of setup.py.

Keep the result (wheel, sdist) to be as much close as possible with the
output of setup.py.

There are some notable changes though:
- werkzeug version has been specified as >=2.0.0, which is where it dropped
  support for python 3.5 and 2.

- wheel now contains the tests and other files as well, similar to sdist.
  These files won't be installed at all, but I haven't found any good way to
  separate the sdist with wheel in terms of the additional files, and I want
  to include these in sdist.

poetry.lock has been also added. This is primairly to have a reproducible
dev environment (eg. mypy, flake8, etc). This will be updated with a bot in
the future.

Fixes #47