doc: fix documentation refereces
There were a couple of warning for "reference target not found" when sphinx
was run with "-n". This has been fixed now.

Intersphinx extensions enabled to reference to external parties, references
to those has been fixed.

api.rst fixed to include the classes which are part of the API (and there
are links to them in the docstrings).

Added "-n" to Makefile to show the reference errors in the future.

Fixes #122