pytest_httpserver: extend `HttpServerBase` to support writing behave test steps in real order
Although HTTPServer is suitable in itself to use in behave tests,
but because of the expectation on a request has to be done
before the request is performed
the test steps of checking the request and sending the response
also has to be before the test step triggering the request.
This makes a confusing test specification like:

Then SUT sends an other request to server
When server responds something
When a request is sent to SUT
Then SUT responds something

Using the BlockingHttpServer, the server blocks until the request is asserted
and then it blocks again until the response is performed.
The test steps can be written in real order:

When a request is sent to SUT
Then SUT sends an other request to server
When server responds something
Then SUT responds something