howto: fix blocking httpserver howto
It is recommended to call check_assertions() at the end of the test run to
check if there were any requests made by the client which were not handled
at all.

In that case the server registers that there was an error and this needs to
be checked at the end.